1. FileGenerationService new methods - added 2. Junit tests for FileGenerationService - added 3. Saving to DB by using postman POST

+++ src/main/java/com/edrone/filegenerator/GeneratorRepository.java
@@ -9,5 +9,5 @@ import org.springframework.stereotype.Service;
 import java.math.BigInteger;
 
 
-public interface GeneratorRepository extends JpaRepository<GeneratedString, BigInteger> {
+public interface GeneratorRepository extends JpaRepository<GeneratedString, Integer> {
 }
